Change id to Integer

+++ src/main/java/example/cashcard/CashCard.java
@@ -2,5 +2,5 @@ package example.cashcard;
 
 import org.springframework.data.annotation.Id;
 
-record CashCard(@Id Long id, Double amount, String owner) {
+record CashCard(@Id Integer id, double amount, String owner) {
 }
